refactor: update around markdown-it plugins
