+++ assets/main.scss
@@ -91,6 +91,29 @@ a {
   }
 }
 
+.link-preview-widget {
+  display: table;
+  width: 100%;
+  padding: 1.2rem 0.8rem;
+  border: 1px solid #e6e6e6;
+  border-radius: 4px;
+}
+
+.link-preview-widget > a {
+  display: table-cell;
+  flex: 1;
+  flex-direction: column;
+  padding: 16px;
+  color: inherit;
+  text-decoration: none;
+  vertical-align: middle;
+  cursor: pointer;
+  background-color: transparent;
+  -webkit-box-orient: vertical;
+  -webkit-box-direction: normal;
+  -webkit-box-flex: 1;
+}
+
 .latest-article {
   display: flex;
   flex-direction: row;
@@ -346,8 +369,18 @@ a {
       }
     }
 
+    .twitter-tweet {
+      width: 100% !important;
+      padding: 10px;
+      background: #fff;
+      border: 1px solid #ccc !important;
+      border-radius: 0 !important;
+      box-shadow: none !important;
+    }
+
     pre {
-      padding: 0.15em 0.25em;
+      padding: 1.2rem 0.8rem;
+      overflow-x: scroll;
       text-align: left;
     }
